rename: 계층명 수정
