+++ src/main/java/chungbazi/chungbazi_be/domain/notification/service/ChatRoomSettingService.java
@@ -5,7 +5,7 @@ import chungbazi.chungbazi_be.domain.chat.service.ChatRoomService;
 import chungbazi.chungbazi_be.domain.chat.entity.ChatRoomSetting;
 import chungbazi.chungbazi_be.domain.chat.repository.ChatRoomSettingRepository;
 import chungbazi.chungbazi_be.domain.user.entity.User;
-import chungbazi.chungbazi_be.domain.user.utils.UserHelper;
+import chungbazi.chungbazi_be.domain.user.support.UserHelper;
 import chungbazi.chungbazi_be.global.apiPayload.code.status.ErrorStatus;
 import chungbazi.chungbazi_be.global.apiPayload.exception.handler.NotFoundHandler;
 import lombok.RequiredArgsConstructor;
