+++ src/main/java/chungbazi/chungbazi_be/domain/user/support/UserHelper.java
@@ -1,5 +1,6 @@
-package chungbazi.chungbazi_be.domain.user.utils;
+package chungbazi.chungbazi_be.domain.user.support;
 
+import chungbazi.chungbazi_be.domain.auth.jwt.SecurityUtils;
 import chungbazi.chungbazi_be.domain.user.entity.User;
 import chungbazi.chungbazi_be.domain.user.repository.UserRepository;
 import chungbazi.chungbazi_be.global.apiPayload.code.status.ErrorStatus;
@@ -9,11 +10,11 @@ import org.springframework.stereotype.Component;
 
 @Component
 @RequiredArgsConstructor
-public class UserReader {
-
+public class UserHelper {
     private final UserRepository userRepository;
 
-    public User getUser(Long userId) {
+    public User getAuthenticatedUser() {
+        Long userId = SecurityUtils.getUserId();
         return userRepository.findById(userId)
                 .orElseThrow(() -> new NotFoundHandler(ErrorStatus.NOT_FOUND_USER));
     }
