--- src/main/java/chungbazi/chungbazi_be/domain/user/utils/UserHelper.java
@@ -1,21 +1,0 @@
-package chungbazi.chungbazi_be.domain.user.utils;
-
-import chungbazi.chungbazi_be.domain.auth.jwt.SecurityUtils;
-import chungbazi.chungbazi_be.domain.user.entity.User;
-import chungbazi.chungbazi_be.domain.user.repository.UserRepository;
-import chungbazi.chungbazi_be.global.apiPayload.code.status.ErrorStatus;
-import chungbazi.chungbazi_be.global.apiPayload.exception.handler.NotFoundHandler;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class UserHelper {
-    private final UserRepository userRepository;
-
-    public User getAuthenticatedUser() {
-        Long userId = SecurityUtils.getUserId();
-        return userRepository.findById(userId)
-                .orElseThrow(() -> new NotFoundHandler(ErrorStatus.NOT_FOUND_USER));
-    }
-}
